chore: use UUID4 as request ID and pass from x-request-id header
The current deployment places the service behind an instance of
an ingress proxy that automatically generates request IDs for
all incoming requests in the form of a UUID4 and provides them
in the `x-request-id` request header, so it makes sense to reuse
the value of this header if present.

Also, the common practice is to use UUID4s for request IDs as
implemented by that proxy, so if the header is not present,
generate a new request ID as a UUID4 rather than an 8-byte
sequence.